+++ server/src/controllers/followers.controller.ts
@@ -1,4 +1,4 @@
-import { DeleteFollowerResponseDTO } from '@/dtos/delete-follower.dto';
+import { DeleteFollowerResponseDTO } from '@/dtos/followers.dto';
 import { deleteFollower } from '@/services/followers.service';
 import { AppError } from '@/utils/app-error';
 import { NextFunction, Request, Response } from 'express';
