--- server/src/dtos/create-post.dto.ts
@@ -1,20 +1,0 @@
-import { z } from 'zod';
-import { createPostSchema } from '@/schemas/post.schema';
-import { TActivity, TMood } from '@/models/post.model';
-
-export type CreatePostRequestDTO = z.infer<typeof createPostSchema>;
-
-export interface CreatePostResponseDTO {
-    status: 'created';
-    post: {
-        id: string;
-        user: {
-            id: string;
-            username: string;
-        };
-        text?: string;
-        activity: TActivity;
-        mood: TMood;
-        createdAt: Date;
-    };
-};
