--- server/src/dtos/delete-comment.dto.ts
@@ -1,4 +1,0 @@
-export interface DeleteCommentResponseDTO {
-    status: 'deleted',
-    message: 'Comment deleted succesfully',
-}
