--- server/src/dtos/delete-post.dto.ts
@@ -1,4 +1,0 @@
-export interface DeletePostResponseDTO {
-    status: 'deleted',
-    message: 'Post deleted successfully',
-};
