--- server/src/dtos/delete-unfollow.dto.ts
@@ -1,4 +1,0 @@
-export interface DeleteUnfollowResponseDTO {
-    status: string;
-    message: string;
-}
