--- server/src/dtos/get-follow-requests.dto.ts
@@ -1,11 +1,0 @@
-export type FollowRequest = {
-    id: string,
-    username: string,
-    bio?: string,
-    isPublic: boolean,
-};
-
-export interface GetFollowRequestsResponseDTO {
-    status: string;
-    followRequests: FollowRequest[];
-};
