--- server/src/dtos/get-post-detail.dto.ts
@@ -1,6 +1,0 @@
-import { GetCommentResponseDTO } from '@/dtos/get-comments.dto';
-import { PostResponseDTO } from '@/dtos/get-post.dto';
-
-export interface GetPostDetailResponseDTO extends PostResponseDTO {
-    comments: GetCommentResponseDTO[];
-};
