--- server/src/dtos/get-post.dto.ts
@@ -1,18 +1,0 @@
-import { TActivity, TMood } from '@/models/post.model';
-
-export interface PostResponseDTO {
-    id: string;
-    user: {
-        id: string;
-        username: string;
-    };
-    text?: string;
-    activity: TActivity;
-    mood: TMood;
-    createdAt: Date;
-};
-
-export type GetPostsResponseDTO = {
-    status: 'ok',
-    posts: PostResponseDTO[],
-};
