--- server/src/dtos/get-profile.dto.ts
@@ -1,12 +1,0 @@
-export interface ProfileResponseDTO {
-    status: 'ok';
-    profile: {
-        username: string;
-        email?: string;
-        birthdate?: string;
-        bio?: string;
-        followers?: number;
-        following?: number;
-        followRequests?: string[];
-    };
-};
