--- server/src/dtos/get-users-followers.dto.ts
@@ -1,18 +1,0 @@
-export type PopulateFollower = {
-    id: string,
-    username: string,
-    bio?: string,
-    followers: string[],
-    following: string[],
-};
-
-export interface GetUsersFollowersResponseDTO {
-    status: string;
-    followers: {
-        id: string,
-        username: string,
-        bio?: string,
-        followers: number,
-        following: number,
-    }[];
-};
