--- server/src/dtos/get-users-following.dto.ts
@@ -1,18 +1,0 @@
-export type PopulateFollowing = {
-    id: string,
-    username: string,
-    bio?: string,
-    followers: string[],
-    following: string[],
-};
-
-export interface GetUsersFollowingResponseDTO {
-    status: string;
-    following: {
-        id: string,
-        username: string,
-        bio?: string,
-        followers: number,
-        following: number,
-    }[];
-};
