--- server/src/dtos/patch-follow-request-accept.dto.ts
@@ -1,10 +1,0 @@
-export interface AcceptFollowRequestResponseDTO {
-    status: 'accepted';
-    message: string;
-    user: {
-        username: string;
-        followers: number;
-        following: number;
-        followRequests: string[];
-    };
-};
