--- server/src/dtos/patch-follow-request-reject.dto.ts
@@ -1,4 +1,0 @@
-export interface RejectFollowRequestResponseDTO {
-    status: 'rejected';
-    message: string; // Follow request from ${username} rejected
-};
