--- server/src/dtos/patch-profile.dto.ts
@@ -1,14 +1,0 @@
-import { patchProfileSchema } from '@/schemas/user.schema';
-import { z } from 'zod';
-
-export type PatchProfileRequestDTO = z.infer<typeof patchProfileSchema>;
-
-export interface PatchProfileResponseDTO {
-    status: 'updated';
-    user: {
-        username: string;
-        email: string;
-        birthdate: string;
-        bio?: string;
-    };
-}
