--- server/src/dtos/post-follow.dto.ts
@@ -1,4 +1,0 @@
-export interface FollowRequestResponseDTO {
-	status: 'followed' | 'pending',
-	message: string,
-}
